+++ src/test/java/org/vaadin/example/addon/AbstractViewTest.java
@@ -1,4 +1,4 @@
-package org.vaadin.example.paperslider;
+package org.vaadin.example.addon;
 
 import org.junit.Assert;
 import org.junit.Before;
