+++ src/test/java/org/vaadin/example/addon/TestViewIT.java
@@ -1,4 +1,4 @@
-package org.vaadin.example.paperslider;
+package org.vaadin.example.addon;
 
 import com.vaadin.flow.component.html.testbench.DivElement;
 import org.junit.Assert;
